Move packet parsing code out of adapter into analyzer
